refactor: rename some modules
