+++ pytorch_lightning/callbacks/pt_callbacks.py
@@ -4,7 +4,7 @@ import logging
 import warnings
 import numpy as np
 
-from pytorch_lightning.pt_overrides.override_data_parallel import LightningDistributedDataParallel
+from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
 
 
 class Callback(object):
