+++ pytorch_lightning/core/memory.py
