+++ pytorch_lightning/testing/__init__.py
@@ -1,6 +1,6 @@
-from .lm_test_module import LightningTestModel
-from .lm_test_module_base import LightningTestModelBase
-from .lm_test_module_mixins import (
+from .test_module import LightningTestModel
+from .test_module_base import LightningTestModelBase
+from .test_module_mixins import (
     LightningValidationStepMixin,
     LightningValidationMixin,
     LightningValidationStepMultipleDataloadersMixin,
