+++ pytorch_lightning/trainer/dp_mixin.py
@@ -1,7 +1,9 @@
 import torch
 
-from pytorch_lightning.pt_overrides.override_data_parallel import (
-    LightningDistributedDataParallel, LightningDataParallel)
+from pytorch_lightning.overrides.data_parallel import (
+    LightningDistributedDataParallel,
+    LightningDataParallel,
+)
 from pytorch_lightning.utilities.debugging import MisconfigurationException
 
 try:
